+++ setup.py
@@ -123,8 +123,8 @@ setup(
     # "scripts" keyword. Entry points provide cross-platform support and allow
     # pip to create the appropriate form of executable for the target platform.
     # entry_points={'console_scripts': ['pyccd-detect=ccd.cli:detect', ], },
-    entry_points = {'console_scripts': ['lcw-listen=cw.__listener__:main',
-                                        'lcw-test-send=cw.__sender__:main']}
+    entry_points = {'console_scripts': ['lcw-listen=cw.__worker_main__:main',
+                                        'lcw-test-send=cw.__test_send__:main']}
     ##entry_points='''
     ##    [core_package.cli_plugins]
     ##    sample=ccd.cli:sample
